+++ src/vs/workbench/contrib/chat/common/chatContext.ts
